discord-invite-rewards-bot/requirements.txt
discord.py
sqlite3
asyncio